+++ cll/model.py
